+++ src/ts/Ads/Models/AuctionResponse.ts
@@ -54,7 +54,6 @@ export interface IRawAuctionV6Response {
 
 export interface IAuctionResponse {
     placements: AuctionPlacement[];
-    auctionId: string;
     contentType: string;
     content: string;
     cacheTTL: number | undefined;
@@ -82,10 +81,9 @@ export interface IAuctionResponse {
 
 export class AuctionResponse extends Model<IAuctionResponse> {
 
-    constructor(placements: AuctionPlacement[], data: IAuctionResponse, mediaId: string, correlationId: string, statusCode?: number, auctionId?: string) {
+    constructor(placements: AuctionPlacement[], data: IAuctionResponse, mediaId: string, correlationId: string, statusCode?: number) {
         super('AuctionResponse', {
             placements: ['array'],
-            auctionId: ['string'],
             contentType: ['string'],
             content: ['string'],
             cacheTTL: ['integer', 'undefined'],
@@ -112,7 +110,6 @@ export class AuctionResponse extends Model<IAuctionResponse> {
         });
 
         this.set('placements', placements);
-        this.set('auctionId', auctionId || ''); // TODO: Make this non-optional
         this.set('contentType', data.contentType);
         this.set('content', data.content);
         this.set('cacheTTL', data.cacheTTL);
@@ -140,10 +137,6 @@ export class AuctionResponse extends Model<IAuctionResponse> {
 
     public getPlacements(): AuctionPlacement[] {
         return this.get('placements');
-    }
-
-    public getAuctionId(): string {
-        return this.get('auctionId');
     }
 
     public getContentType(): string {