+++ src/ts/Ads/Networking/AuctionRequest.ts
@@ -476,7 +476,6 @@ export class AuctionRequest {
                     omidJSVersion: OM_JS_VERSION,
                     legalFramework: this._privacySDK.getLegalFramework(),
                     agreedOverAgeLimit: this._userPrivacyManager.getAgeGateChoice()
-                    //Todo: Add IsMadeWithUnity flag from SDKDetectionInfo
                 };
             });
         });